# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto Backend
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can
    upload your photos directly from your PC, and they will be visible to everyone following you.
  contact:
    email: deodati.2025953@studenti.uniroma1.it
  license:
    name: GPL 3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 1.0.0
servers:
  - url: "http://localhost:3000"
...
paths:
...
  /users/me/username:
    put:
      tags:
        - Account management
      summary: Sets the username
      description: Changes the logged in user's username. Returns the same username on success
      operationId: setMyUserName
      requestBody:
        description: Username to set
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Username'
      responses:
        '200':
          description: Operation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Username'
        '400':
          description: Username already taken
        '401':
          description: Unauthenticated
  /users/me/pp:
    put:
      tags:
        - Account management
      summary: Set a profile picture
      description: |-
        Sets the profile picture for the logged in user, the picture should be provided in
        the request body.
      operationId: setProPic
      requestBody:
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '204':
          description: Operation successful
        '400':
          description: Bad image
        '401':
          description: Unauthenticated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Username:
      type: string
      example: Maria
      pattern: 'ˆ.*?$'
      description: Case-insensitive username
      minLength: 3
      maxLength: 40
...
security:
  - bearerAuth: []